requirements.txt
# Backend
fastapi
uvicorn[standard]

# Frontend
streamlit

# Services & Data
requests
pandas
openai
python-dotenv
beautifulsoup4
lxml
